requirements.txt
netifaces=0.10.9
CoAPthon3==1.0.1